[UPDATE] Building Cart: Update and Remove item in cart

+++ Projects/public/js/20b01.js
@@ -31,7 +31,11 @@ function removeCartItem(id) {
         success: function(result) {
             $('#cart-badge').html(result.totalQuantity);
             $('#totalPrice').html('$' + result.totalPrice);
-            $(`#item${id}`).remove();
+            if (result.totalQuantity > 0) {
+                $(`#item${id}`).remove();
+            } else {
+                $('#cart-body').html('<div class="alert alert-info text-center">Your cart is empty!</div>')
+            }
         }
     });
 }
@@ -50,13 +54,14 @@ function updateCartItem(id, quantity) {
 }
 
 function clearCart() {
-    $.ajax({
-        url: '/cart/all',
-        type: 'DELETE',
-        data: { id },
-        success: function() {
-            $('#cart-badge').html('0');
-            $('#cart-body').html('<div class="alert alert-info text-center">Ypur cart is empty!</div>')
-        }
-    });
+    if (confirm('Do you really want to remove all items?')) {
+        $.ajax({
+            url: '/cart/all',
+            type: 'DELETE',
+            success: function() {
+                $('#cart-badge').html('0');
+                $('#cart-body').html('<div class="alert alert-info text-center">Your cart is empty!</div>')
+            }
+        });
+    }
 }
